ytsentiment/requirements.txt

pandas
matplotlib
seaborn
google-api-python-client
nltk
python-dotenv
